+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.0.7
+TBC: Add interrupt duration lockout fix for a few 20% reduction items.
+TBC: Remove wyvern sting dot spell id
+
 # 1.0.6
 Add dubious TBC support