+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
-# 1.0.6
-Add dubious TBC support
+# 1.0.7
+TBC: Add interrupt duration lockout fix for a few 20% reduction items.
+TBC: Remove wyvern sting dot spell id